test(shots): consolidate FAB suppression pins into the pre-existing test file
CI round 1 caught what my targeted run missed: FloatingActionBar already had a
component-adjacent test file (not under __tests__/) whose un-mocked renders
broke when the component gained the useResolvedSurface call. The 5 existing
route/device pins now run with steerable surface/flag mocks (all behaviors
unchanged), my duplicate __tests__ file is deleted, and the 3 suppression pins
live alongside them — 8/8 green.
