+++ src-vnext/shared/components/FloatingActionBar.test.tsx
@@ -4,6 +4,10 @@ import { MemoryRouter } from "react-router-dom"
 import { FloatingActionBar } from "./FloatingActionBar"
 
 const mockUseIsDesktop = vi.fn(() => false)
+const mocks = vi.hoisted(() => ({
+  surface: "plan-build" as string | null,
+  shootFlag: false,
+}))
 
 vi.mock("@/shared/hooks/useMediaQuery", () => ({
   useIsMobile: () => !mockUseIsDesktop(),
@@ -11,6 +15,30 @@ vi.mock("@/shared/hooks/useMediaQuery", () => ({
   useIsTablet: () => false,
   useMediaQuery: () => false,
 }))
+
+// The component consults the resolved surface only to suppress itself on the
+// Shoot shell (5e-II eyeball finding: the FAB overlapped the shell's sticky
+// bottom status bar). Mock the hook directly — this chrome unit test renders
+// without auth/role providers.
+vi.mock("@/features/shots/hooks/useResolvedSurface", () => ({
+  useResolvedSurface: () => ({
+    surface: mocks.surface,
+    affordances: null,
+    chrome: null,
+    resolving: false,
+  }),
+}))
+
+vi.mock("@/shared/lib/flags", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("@/shared/lib/flags")>()
+  return {
+    ...actual,
+    isFeatureEnabled: (flag: string) =>
+      flag === "featureShootSurface"
+        ? mocks.shootFlag
+        : actual.isFeatureEnabled(flag as never),
+  }
+})
 
 function renderFab(route: string) {
   return render(
@@ -23,6 +51,8 @@ function renderFab(route: string) {
 describe("FloatingActionBar", () => {
   beforeEach(() => {
     mockUseIsDesktop.mockReturnValue(false)
+    mocks.surface = "plan-build"
+    mocks.shootFlag = false
   })
 
   it("renders on shot list route for mobile", () => {
@@ -50,5 +80,26 @@ describe("FloatingActionBar", () => {
   it("does NOT render on pull routes", () => {
     renderFab("/projects/p1/pulls")
     expect(screen.queryByTestId("fab")).not.toBeInTheDocument()
+  })
+
+  describe("Shoot shell suppression (5e-II)", () => {
+    it("suppresses itself on the Shoot surface (flag on)", () => {
+      mocks.shootFlag = true
+      mocks.surface = "shoot"
+      renderFab("/projects/p1/shots")
+      expect(screen.queryByTestId("fab")).not.toBeInTheDocument()
+    })
+
+    it("still renders for plan-build surfaces with the flag on", () => {
+      mocks.shootFlag = true
+      renderFab("/projects/p1/shots")
+      expect(screen.getByTestId("fab")).toBeInTheDocument()
+    })
+
+    it("does NOT suppress while the flag is off, even if the surface resolves shoot", () => {
+      mocks.surface = "shoot"
+      renderFab("/projects/p1/shots")
+      expect(screen.getByTestId("fab")).toBeInTheDocument()
+    })
   })
 })
